🦄 refactor(home-layout): refactoring layout structure
