+++ src/data/index.ts
@@ -1,0 +1,2 @@
+export * from './sidebar-items'
+export * from './profile-menu-items'
