+++ src/presentations/layouts/protected/home/index.ts
@@ -1,1 +1,4 @@
 export * from './home'
+export * from './sidebar'
+export * from './menu-item'
+export * from './mobile-nav'
