+++ src/presentations/layouts/protected/home/menu-item.tsx
@@ -1,0 +1,41 @@
+import { Flex, FlexProps, Icon } from "@chakra-ui/react";
+import { IconType } from "react-icons";
+import { Link, To } from "react-router-dom";
+
+
+interface MenuItemProps extends FlexProps {
+  icon?: IconType;
+  to: To
+  children: string;
+}
+
+export const MenuItem = ({ icon, to, children, ...rest }: MenuItemProps) => {
+  return (
+    <Link to={to} style={{ textDecoration: 'none' }}>
+      <Flex
+        align="center"
+        p="4"
+        mx="4"
+        borderRadius="lg"
+        role="group"
+        cursor="pointer"
+        _hover={{
+          bg: 'cyan.400',
+          color: 'white',
+        }}
+        {...rest}>
+        {icon && (
+          <Icon
+            mr="4"
+            fontSize="16"
+            _groupHover={{
+              color: 'white',
+            }}
+            as={icon}
+          />
+        )}
+        {children}
+      </Flex>
+    </Link>
+  );
+};
