+++ src/stores/slices/index.ts
@@ -1,2 +1,3 @@
 export * from './auth-slice'
 export * from './user-slice'
+export * from './sidebar-slice'
